Updated User Interface and Functionality
I got rid of the majority of commands and restructured the code over and
over till the point where it all works. I didn't know async and await
were a thing.
Okay so first thing we use regex to match the autocomplete suggestions,
so now the suggestion box works as intended. In addition, the majority
of commands have been removed because ... they weren't fit for the
omnibox interface. I suppose that should've been noted. Eventually I
might add a settings page with editable options, but no point for now.