Added Popup for WD List
This does not complete the WD List implementation. It will handle cases
where a webpage has already been loaded. At present, it requires
permissions to all websites so it can create its popup.

This commit adds all the styling, scripting, and html needed for
building a popup and potentially a extension page in the future. In
addition, in the future the host permission should be made optional and
requested on the first user interaction with wd list.

The popup at present simply displays every potentially enterable
shortcut the user has defined and can click at any moment.